Noting issues with CLUTTER-READ-DELIMITED-LIST
There's a bug, demonstrated by a test case, where C-R-D-L fails when end-chars that are not already
associated with a reader macro are used. I've made C-R-D-L actually die with an EOF when errors such
as this happen (instead of doing something weird with NIL or signaling a type error because of
CHAR=), and I've annotated the cause of the bug in the test case.